refactor: 패키지 위치 조정
