+++ Moyoy-Api/src/main/java/com/moyoy/api/auth/security/component/CustomAuthenticationEntryPoint.java
@@ -12,7 +12,7 @@ import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 import com.moyoy.api.auth.error.AuthErrorCode;
-import com.moyoy.api.common.util.ErrorResponseWriter;
+import com.moyoy.api.support.ErrorResponseWriter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
