+++ Moyoy-Api/src/main/java/com/moyoy/api/auth/security/component/HttpCookieOAuth2AuthorizationRequestRepository.java
@@ -4,7 +4,7 @@ import org.springframework.security.oauth2.client.web.AuthorizationRequestReposi
 import org.springframework.security.oauth2.core.endpoint.OAuth2AuthorizationRequest;
 import org.springframework.stereotype.Component;
 
-import com.moyoy.api.common.util.CookieUtils;
+import com.moyoy.api.support.CookieUtils;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
