+++ Moyoy-Api/src/main/java/com/moyoy/api/github_follow/application/GithubFollowDetectService.java
@@ -9,7 +9,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 
-import com.moyoy.api.common.util.SlicePagingUtils;
+import com.moyoy.api.support.SlicePagingUtils;
 import com.moyoy.api.github_follow.application.event.DetectEvent;
 import com.moyoy.api.github_follow.application.request.GithubFollowDetectionData;
 import com.moyoy.api.github_follow.application.response.GithubFollowDetectionResult;
